configs: bump zedboard to 2014.1
Bump zedboard defconfig to version 2014.1.  This makes it possible to use
uboot-spl instead of the xilinx specific fsbl.  However, the result is kinda
hacky due to the licensing state of ps7_init.{c,h} needed for building uboot
spl.  Directions for building a proper boot.bin using the new methodolgy is
included in the readme.txt.